Move nav search shortcut test to dedicated spec
